solve a strange mismatch error

+++ src/padding.rs
@@ -26,7 +26,7 @@ use std::convert::TryInto;
 /// The function returns `Err(())` if the padded buffer length would exceed `max_buflen`, or if the
 /// block size is 0. It returns a result containing the new padded length upon success.
 pub fn pad(buf: &mut [u8], unpadded_buflen: usize, blocksize: usize) -> Result<usize, ()> {
-    let mut padded_buflen_p: u64 = 0;
+    let mut padded_buflen_p: u32 = 0;
     unsafe {
         if 0 == ffi::sodium_pad(
             &mut padded_buflen_p,
@@ -45,7 +45,7 @@ pub fn pad(buf: &mut [u8], unpadded_buflen: usize, blocksize: usize) -> Result<u
 /// The `unpad()` function computes the original, unpadded length of a message previously padded
 /// using [`pad()`]. The original length is returned upon success.
 pub fn unpad(buf: &[u8], padded_buflen: usize, blocksize: usize) -> Result<usize, ()> {
-    let mut unpadded_buflen_p: u64 = 0;
+    let mut unpadded_buflen_p: u32 = 0;
     unsafe {
         if 0 == ffi::sodium_unpad(
             &mut unpadded_buflen_p,
